MEDIC-50: Поменять дизайн карточек аптечек
- Заменил дизайн ячейки карточки аптечки
- Добавил новые изображения
- Удалил старые классы ячеек
- Поменял диаметр изображений у кнопок действий по свайпу
