+++ MedicinesV2/Default/Recources/Localizable.xcstrings
@@ -11,10 +11,19 @@
         }
       }
     },
+    "Лекарств в аптечке" : {
+
+    },
     "Не задано" : {
 
     },
     "Остаталось" : {
+
+    },
+    "Открыть" : {
+
+    },
+    "Просрочено" : {
 
     },
     "Редактировать" : {
