+++ MedicinesV2/PresentationLayer/MainScreen/Container/FirstAidKits/Cells/FirstAidKitCell.swift
@@ -2,50 +2,175 @@
 //  FirstAidKitCell.swift
 //  MedicinesV2
 //
-//  Created by Дмитрий Данилин on 18.10.2022.
+//  Created by Дмитрий Данилин on 09.10.2023.
 //
 
 import UIKit
 
-/// Ячейка для аптечки
+/// Ячейка для лекарства
 final class FirstAidKitCell: UITableViewCell, IdentifiableCell {
     
-    // MARK: - Private properties
+    /// Замыкание, которое срабатывает после нажатия на кнопку ``pressedEditButton``
+    /// - Нужно для того, чтобы перейти на следующий экран с лекарствами в аптечке.
+    var buttonTappedAction: (() -> Void)?
     
-    /// Кастомный контейнер
-    /// - В нем содержаться все элементы аптечки
-    /// - нужен для того, чтобы сделать его в виде карточки
-    private let viewContainer: UIView = {
-        let view = UIView()
+    // MARK: - Private properties
+
+    private let containerVStackView: UIStackView = {
+        let view = UIStackView()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.backgroundColor = .systemBackground
-        view.layer.cornerRadius = 16
+        view.axis = .vertical
+        view.spacing = Indents.edgeMarginPadding
+        view.layer.cornerRadius = 30
+        view.isLayoutMarginsRelativeArrangement = true
+        view.layoutMargins = UIEdgeInsets(
+            top: Indents.edgeMarginPadding,
+            left: Indents.edgeMarginPadding,
+            bottom: Indents.edgeMarginPadding,
+            right: Indents.edgeMarginPadding
+        )
         return view
     }()
     
-    private let titleFirstAidKitLabel: UILabel = {
+    private let titleContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+    
+    private let nameLabel: UILabel = {
         let label = UILabel()
-        label.font = .systemFont(ofSize: 17)
-        label.numberOfLines = 0
-        label.setContentCompressionResistancePriority(.defaultLow, for: .vertical)
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = Fonts.systemNormal(.size1).font
+        label.numberOfLines = 2
         return label
     }()
     
-    private let expiredMedicinesCountLabel: UILabel = {
+    private let iconImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.backgroundColor = .systemGray6
+        imageView.layer.cornerRadius = Constants.iconMedecineSize / 2
+        imageView.contentMode = .scaleAspectFit
+        imageView.clipsToBounds = true
+        return imageView
+    }()
+    
+    private let separatorTopView: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = .systemGray6
+        return view
+    }()
+    
+    private let descriptionLabel: UILabel = {
         let label = UILabel()
-        label.font = .systemFont(ofSize: 12)
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = Fonts.systemNormal(.size2).font
+        label.textColor = .systemGray
+        label.numberOfLines = 2
         return label
     }()
     
-    private let amountMedicinesLabel: UILabel = {
+    private let calculableContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.alignment = .top
+        view.distribution = .fillEqually
+        return view
+    }()
+    
+    private let amountContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.axis = .vertical
+        view.alignment = .leading
+        view.spacing = Indents.elementsPadding
+        return view
+    }()
+    
+    private let staticAmountLabel: UILabel = {
         let label = UILabel()
-        label.font = .systemFont(ofSize: 17)
-        label.textColor = .lightGray
-        label.setContentHuggingPriority(.defaultHigh, for: .horizontal)
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = String(localized: "Лекарств в аптечке")
+        label.font = Fonts.systemNormal(.size3).font
         return label
+    }()
+    
+    private let amountLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = Fonts.systemNormal(.size2).font
+        return label
+    }()
+    
+    private let expiryDateContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.axis = .vertical
+        view.alignment = .trailing
+        view.spacing = Indents.elementsPadding
+        return view
+    }()
+    
+    private let staticExpiredAmountLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = String(localized: "Просрочено")
+        label.font = Fonts.systemNormal(.size3).font
+        return label
+    }()
+    
+    private let expiredAmountLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = Fonts.systemNormal(.size2).font
+        return label
+    }()
+    
+    private let separatorBottomView: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = .systemGray6
+        return view
+    }()
+    
+    private let bottomContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.distribution = .equalSpacing
+        return view
+    }()
+    
+    /// Иконка действия
+    /// - показывает иконку, о необходимости обратить внимание на лекарства в аптечке:
+    private let alertIcon: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.contentMode = .scaleAspectFit
+        imageView.image = SystemIcons.alert.image
+        imageView.tintColor = Colors.pinkRed
+        return imageView
+    }()
+    
+    private lazy var openButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.translatesAutoresizingMaskIntoConstraints = false
+        button.setTitle(String(localized: "Открыть"), for: .normal)
+        button.titleLabel?.font = Fonts.systemNormal(.size2).font
+        button.setTitleColor(.white, for: .normal)
+        button.backgroundColor = Colors.darkCyan
+        button.layer.cornerRadius = 15
+        button.accessibilityLabel = "Open Button"
+        
+        let tapAction = UIAction(title: "tap action") { [weak self] _ in
+            self?.pressedEditButton()
+        }
+        
+        button.addAction(tapAction, for: .touchUpInside)
+        
+        return button
     }()
     
     // MARK: - Initializer
@@ -64,6 +189,19 @@ final class FirstAidKitCell: UITableViewCell, IdentifiableCell {
     
     // MARK: - Override Methods
     
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        
+        nameLabel.text = ""
+        iconImageView.image = nil
+        descriptionLabel.isHidden = false
+        descriptionLabel.text = ""
+        amountLabel.text = ""
+        expiredAmountLabel.text = ""
+        
+        setAlertsDefault()
+    }
+    
     override func traitCollectionDidChange(
         _ previousTraitCollection: UITraitCollection?
     ) {
@@ -72,42 +210,69 @@ final class FirstAidKitCell: UITableViewCell, IdentifiableCell {
         setupCardShadowColor()
     }
     
-}
-
-// MARK: - Publik Methods
-
-extension FirstAidKitCell {
-    /// Метод для конфигурирования отображаемой информации в ячейке.
-    /// - Parameters:
-    ///   - titleFirstAidKit: принимает название аптечки.
-    ///   - amountMedicines: принимает количество лекарств, хранящихся  в аптечке.
-    func configure(titleFirstAidKit: String?, amountMedicines: String, expiredCount: Int) {
-        self.titleFirstAidKitLabel.text = titleFirstAidKit
-        self.amountMedicinesLabel.text = amountMedicines
-        
-        if expiredCount != 0 {
-            expiredMedicinesCountLabel.text = "Просроченных лекарств: \(expiredCount)"
-            expiredMedicinesCountLabel.textColor = Colors.pinkRed
-        } else {
-            expiredMedicinesCountLabel.text = "Все лекарства в порядке"
-            expiredMedicinesCountLabel.textColor = Colors.lightGreen
-        }
-        
+    // MARK: - Actions
+    
+    private func pressedEditButton() {
+        buttonTappedAction?()
     }
 }
 
-// MARK: - Конфигурирование ячейки
+// MARK: - Public methods
+
+extension FirstAidKitCell {
+    func configure(
+        image: String? = nil,
+        name: String,
+        description: String? = nil,
+        expiredAmount: Int?,
+        amount: Int?
+    ) {
+        setImage(from: image)
+        nameLabel.text = name
+        
+        if description == "" || description == nil {
+            descriptionLabel.isHidden = true
+        } else {
+            descriptionLabel.text = description
+        }
+        
+        amountLabel.text = "\(amount ?? 0) шт"
+        
+        expiredAmountLabel.text = "\(expiredAmount ?? 0) шт"
+        if expiredAmount != 0 {
+            alertIcon.isHidden = false
+            expiredAmountLabel.textColor = Colors.pinkRed
+        } else {
+            setAlertsDefault()
+        }
+    }
+    
+    // TODO: (Задача не заведена) Будет добавлена установка изображений пользователем
+    // к примеру выбор типа расположения аптечки. И соответствующая иконка
+    private func setImage(from url: String?) {
+        iconImageView.image = SystemIcons.firstAidKit.image
+        iconImageView.tintColor = Colors.lightGreen
+        iconImageView.frame = CGRect(x: 0, y: 0, width: 20, height: 20)
+    }
+    
+    /// Метод для сброса всех предупреждений
+    /// - используется для сброса расцветки лейблов и иконки в случае их изменения, а так же при
+    /// переиспользовании ячеек, к примеру в процессе поиска.
+    private func setAlertsDefault() {
+        expiredAmountLabel.textColor = .label
+        alertIcon.isHidden = true
+    }
+}
+
+// MARK: - Configuration methods
 
 private extension FirstAidKitCell {
-    /// Метод инициализации настроек ячейки
     func setup() {
         setupUI()
     }
     
-    /// Метод для настройки отображения элементов
     func setupUI() {
         selectionStyle = .none
-        contentView.backgroundColor = .clear
         backgroundColor = .clear
         
         setupCardShadow()
@@ -116,9 +281,9 @@ private extension FirstAidKitCell {
     /// Метод настройки параметров тени карточки
     func setupCardShadow() {
         setupCardShadowColor()
-        viewContainer.layer.shadowOffset = CGSize(width: 0, height: 3)
-        viewContainer.layer.shadowRadius = 3
-        viewContainer.layer.shadowOpacity = 0.5
+        containerVStackView.layer.shadowOffset = CGSize(width: 0, height: 5)
+        containerVStackView.layer.shadowRadius = 5
+        containerVStackView.layer.shadowOpacity = 0.3
     }
     
     /// Используется для изменения цвета тени карточки лекарства
@@ -126,33 +291,67 @@ private extension FirstAidKitCell {
     /// нужно вызывать в traitCollectionDidChange для изменение цвета. Так как cgColor
     /// не изменяются сами.
     func setupCardShadowColor() {
-        viewContainer.layer.shadowColor = Colors.cardShadow.cgColor
+        containerVStackView.layer.shadowColor = Colors.cardShadow.cgColor
     }
     
     func addViews() {
-        contentView.addSubview(viewContainer)
-        viewContainer.addSubview(titleFirstAidKitLabel)
-        viewContainer.addSubview(expiredMedicinesCountLabel)
-        viewContainer.addSubview(amountMedicinesLabel)
+        contentView.addSubview(containerVStackView)
+        
+        containerVStackView.addArrangedSubview(titleContentStackView)
+        titleContentStackView.addArrangedSubview(nameLabel)
+        titleContentStackView.addArrangedSubview(iconImageView)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: titleContentStackView)
+        containerVStackView.addArrangedSubview(separatorTopView)
+        
+        containerVStackView.addArrangedSubview(descriptionLabel)
+        
+        containerVStackView.addArrangedSubview(calculableContentStackView)
+        calculableContentStackView.addArrangedSubview(amountContentStackView)
+        amountContentStackView.addArrangedSubview(staticAmountLabel)
+        amountContentStackView.addArrangedSubview(amountLabel)
+        
+        calculableContentStackView.addArrangedSubview(expiryDateContentStackView)
+        expiryDateContentStackView.addArrangedSubview(staticExpiredAmountLabel)
+        expiryDateContentStackView.addArrangedSubview(expiredAmountLabel)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: calculableContentStackView)
+        containerVStackView.addArrangedSubview(separatorBottomView)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: separatorBottomView)
+        containerVStackView.addArrangedSubview(bottomContentStackView)
+        bottomContentStackView.addArrangedSubview(alertIcon)
+        bottomContentStackView.addArrangedSubview(UIView())
+        bottomContentStackView.addArrangedSubview(openButton)
     }
     
     func setupConstraints() {
         NSLayoutConstraint.activate([
-            viewContainer.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-            viewContainer.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            viewContainer.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            viewContainer.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
+            containerVStackView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: Indents.edgeMarginPadding),
+            containerVStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: Indents.edgeMarginPadding),
+            containerVStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -Indents.edgeMarginPadding),
+            containerVStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -Indents.edgeMarginPadding),
+                        
+            iconImageView.heightAnchor.constraint(equalToConstant: Constants.iconMedecineSize),
+            iconImageView.widthAnchor.constraint(equalToConstant: Constants.iconMedecineSize),
             
-            titleFirstAidKitLabel.topAnchor.constraint(equalTo: viewContainer.topAnchor, constant: 10),
-            titleFirstAidKitLabel.leadingAnchor.constraint(equalTo: viewContainer.leadingAnchor, constant: 16),
-            titleFirstAidKitLabel.trailingAnchor.constraint(equalTo: amountMedicinesLabel.leadingAnchor, constant: -5),
+            separatorTopView.heightAnchor.constraint(equalToConstant: Constants.heightBorder),
+            separatorBottomView.heightAnchor.constraint(equalToConstant: Constants.heightBorder),
             
-            expiredMedicinesCountLabel.topAnchor.constraint(equalTo: titleFirstAidKitLabel.bottomAnchor, constant: 6),
-            expiredMedicinesCountLabel.leadingAnchor.constraint(equalTo: viewContainer.leadingAnchor, constant: 16),
-            expiredMedicinesCountLabel.bottomAnchor.constraint(equalTo: viewContainer.bottomAnchor, constant: -10),
+            alertIcon.widthAnchor.constraint(equalToConstant: Constants.iconActionSize),
             
-            amountMedicinesLabel.centerYAnchor.constraint(equalTo: viewContainer.centerYAnchor),
-            amountMedicinesLabel.trailingAnchor.constraint(equalTo: viewContainer.trailingAnchor, constant: -16)
+            openButton.heightAnchor.constraint(equalToConstant: Constants.editButtonHeight),
+            openButton.widthAnchor.constraint(equalToConstant: Constants.editBottonWidth)
         ])
+    }
+}
+
+private extension FirstAidKitCell {
+    struct Constants {
+        static let iconMedecineSize: CGFloat = 48
+        static let iconActionSize: CGFloat = 28
+        static let heightBorder: CGFloat = 2
+        static let editButtonHeight: CGFloat = 50
+        static let editBottonWidth: CGFloat = UIScreen.main.bounds.width * 0.4139 // примерно 178
     }
 }
