+++ MedicinesV2/PresentationLayer/Medicines/Cells/MedicineCell.swift
@@ -2,60 +2,146 @@
 //  MedicineCell.swift
 //  MedicinesV2
 //
-//  Created by Дмитрий Данилин on 19.10.2022.
+//  Created by Дмитрий Данилин on 03.10.2023.
 //
 
 import UIKit
+import DTLogger
 
 /// Ячейка для лекарства
 final class MedicineCell: UITableViewCell, IdentifiableCell {
     
-    // MARK: - Private properties
+    /// Замыкание, которое срабатывает после нажатия на кнопку ``pressedEditButton``
+    /// - Нужно для того, чтобы перейти на следующий экран редактирования лекарства.
+    var buttonTappedAction: (() -> Void)?
     
-    /// Кастомный контейнер
-    /// - В нем содержаться все элементы лекарства
-    /// - нужен для того, чтобы сделать его в виде карточки
-    private let viewContainer: UIView = {
-        let view = UIView()
+    // MARK: - Private properties
+
+    private let containerVStackView: UIStackView = {
+        let view = UIStackView()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.backgroundColor = .systemBackground
-        view.layer.cornerRadius = 16
+        view.axis = .vertical
+        view.spacing = Indents.edgeMarginPadding
+        view.layer.cornerRadius = 30
+        view.isLayoutMarginsRelativeArrangement = true
+        view.layoutMargins = UIEdgeInsets(
+            top: Indents.edgeMarginPadding,
+            left: Indents.edgeMarginPadding,
+            bottom: Indents.edgeMarginPadding,
+            right: Indents.edgeMarginPadding
+        )
         return view
     }()
     
-    /// Стек лейблов
-    /// - Предназначен для: названия лекарства, типа лекарства, назначения, срока годности
-    private var stackMedicineLabels: UIStackView = {
-        let stack = UIStackView()
-        stack.translatesAutoresizingMaskIntoConstraints = false
-        stack.axis = .vertical
-        stack.distribution = .equalSpacing
-        stack.spacing = 4
-        return stack
+    private let titleContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
     }()
     
     private let nameLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont(name: "HelveticaNeue", size: 20)
+        label.font = Fonts.systemNormal(.size1).font
         label.numberOfLines = 2
         return label
+    }()
+    
+    private let iconImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        imageView.backgroundColor = .systemGray6
+        imageView.layer.cornerRadius = Constants.iconMedecineSize / 2
+        imageView.contentMode = .scaleAspectFit
+        imageView.clipsToBounds = true
+        return imageView
+    }()
+    
+    private let separatorTopView: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = .systemGray6
+        return view
     }()
     
     private let descriptionLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont(name: "HelveticaNeue", size: 15)
-        label.textColor = .systemGray3
+        label.font = Fonts.systemNormal(.size2).font
+        label.textColor = .systemGray
         label.numberOfLines = 2
+        return label
+    }()
+    
+    private let calculableContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.alignment = .top
+        view.distribution = .fillEqually
+        return view
+    }()
+    
+    private let amountContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.axis = .vertical
+        view.alignment = .leading
+        view.spacing = Indents.elementsPadding
+        return view
+    }()
+    
+    private let staticAmountLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = String(localized: "Остаталось")
+        label.font = Fonts.systemNormal(.size3).font
+        return label
+    }()
+    
+    private let amountLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = Fonts.systemNormal(.size2).font
+        return label
+    }()
+    
+    private let expiryDateContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.axis = .vertical
+        view.alignment = .trailing
+        view.spacing = Indents.elementsPadding
+        return view
+    }()
+    
+    private let staticExpiryDateLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.text = String(localized: "Срок годности")
+        label.font = Fonts.systemNormal(.size3).font
         return label
     }()
     
     private let expiryDateLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont(name: "HelveticaNeue", size: 15)
+        label.font = Fonts.systemNormal(.size2).font
         return label
+    }()
+    
+    private let separatorBottomView: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = .systemGray6
+        return view
+    }()
+    
+    private let bottomContentStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.distribution = .equalSpacing
+        return view
     }()
     
     /// Иконка действия
@@ -69,11 +155,23 @@ final class MedicineCell: UITableViewCell, IdentifiableCell {
         return imageView
     }()
     
-    private let amountLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = UIFont(name: "HelveticaNeue", size: 15)
-        return label
+    private lazy var editButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.translatesAutoresizingMaskIntoConstraints = false
+        button.setTitle(String(localized: "Редактировать"), for: .normal)
+        button.titleLabel?.font = Fonts.systemNormal(.size2).font
+        button.setTitleColor(.white, for: .normal)
+        button.backgroundColor = Colors.darkCyan
+        button.layer.cornerRadius = 15
+        button.accessibilityLabel = "Open Button"
+        
+        let tapAction = UIAction(title: "tap action") { [weak self] _ in
+            self?.pressedEditButton()
+        }
+        
+        button.addAction(tapAction, for: .touchUpInside)
+        
+        return button
     }()
     
     // MARK: - Initializer
@@ -92,6 +190,19 @@ final class MedicineCell: UITableViewCell, IdentifiableCell {
     
     // MARK: - Override Methods
     
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        
+        nameLabel.text = ""
+        iconImageView.image = nil
+        descriptionLabel.isHidden = false
+        descriptionLabel.text = ""
+        amountLabel.text = ""
+        expiryDateLabel.text = ""
+        
+        setImageActionIconDefault()
+    }
+    
     override func traitCollectionDidChange(
         _ previousTraitCollection: UITraitCollection?
     ) {
@@ -100,29 +211,42 @@ final class MedicineCell: UITableViewCell, IdentifiableCell {
         setupCardShadowColor()
     }
     
-    // MARK: - Public method
+    // MARK: - Actions
     
-    /// Настройка информации, которая будет отображаться в ячейке лекарства
-    /// - Parameters:
-    ///   - name: принимает название лекарства
-    ///   - type: принимает тип лекарства
-    ///   - expiryDate: принимает дату срока годности лекарства
-    ///   - amount: принимает количество оставшегося лекарства
-    func configure (
+    private func pressedEditButton() {
+        buttonTappedAction?()
+    }
+}
+
+// MARK: - Public methods
+
+extension MedicineCell {
+    func configure(
+        image: String? = nil,
         name: String,
         type: String?,
         purpose: String?,
         expiryDate: Date?,
         amount: NSNumber?
     ) {
-        let description = generateDescriptionFrom(type, purpose)
-        
+        setImage(from: image)
         nameLabel.text = name
+        
+        let description = generateDescriptionFrom(type, purpose)
         descriptionLabel.text = description
-        expiryDateLabel.text = expiryDate?.toString()
+        
         amountLabel.text = "\(amount ?? 0) шт"
         
-        setImageActionIconDefault()
+        if let expiryDate {
+            expiryDateLabel.text = expiryDate.toString()
+        } else {
+            expiryDateLabel.text = String(localized: "Не задано")
+            expiryDateLabel.textColor = Colors.pinkRed
+        }
+                
+        if description == "" {
+            descriptionLabel.isHidden = true
+        }
         
         if let expiryDate, Date() >= expiryDate {
             setImageActionIcon(need: .thrownOut)
@@ -131,24 +255,12 @@ final class MedicineCell: UITableViewCell, IdentifiableCell {
         }
     }
     
-    // MARK: - Private method
-    
-    /// Метод генерации описания лекарсва из типа и назначения
-    /// - Parameters:
-    ///   - type: принимает тип лекарства.
-    ///   - purpose: принимает назначение (область применения) лекарства.
-    /// - Returns: возвращает сгенерированное описание.
-    /// - Используется для правильной генерации строки описания, с запятыми и пробелами
-    private func generateDescriptionFrom(_ type: String?, _ purpose: String?) -> String? {
-        if let type, type != "", let purpose, purpose != "" {
-            return "\(type), \(purpose)"
-        } else if let type, type != "" {
-            return type
-        } else if let purpose {
-            return purpose
-        } else {
-            return nil
-        }
+    // TODO: (Следующий спринт) Будет добавлена установка изображений. 
+    // Пока думаю, брать из сети, или давать это делать пользователю.
+    // Либо встроить иконки в приложение
+    private func setImage(from url: String?) {
+        iconImageView.image = SystemIcons.pills.image
+        iconImageView.tintColor = Colors.lightGreen
     }
     
     /// Настройка иконки предупреждения
@@ -175,21 +287,35 @@ final class MedicineCell: UITableViewCell, IdentifiableCell {
         expiryDateLabel.textColor = .label
         amountLabel.textColor = .label
     }
+    
+    /// Метод генерации описания лекарсва из типа и назначения
+    /// - Parameters:
+    ///   - type: принимает тип лекарства.
+    ///   - purpose: принимает назначение (область применения) лекарства.
+    /// - Returns: возвращает сгенерированное описание.
+    /// - Используется для правильной генерации строки описания, с запятыми и пробелами
+    private func generateDescriptionFrom(_ type: String?, _ purpose: String?) -> String? {
+        if let type, type != "", let purpose, purpose != "" {
+            return "\(type), \(purpose)"
+        } else if let type, type != "" {
+            return type
+        } else if let purpose {
+            return purpose
+        } else {
+            return nil
+        }
+    }
 }
 
-// MARK: - Конфигурирование ячейки
+// MARK: - Configuration methods
 
 private extension MedicineCell {
-
-    /// Метод инициализации настроек ячейки
     func setup() {
         setupUI()
     }
     
-    /// Метод для настройки отображения элементов
     func setupUI() {
         selectionStyle = .none
-        contentView.backgroundColor = .clear
         backgroundColor = .clear
         
         setupCardShadow()
@@ -198,9 +324,9 @@ private extension MedicineCell {
     /// Метод настройки параметров тени карточки
     func setupCardShadow() {
         setupCardShadowColor()
-        viewContainer.layer.shadowOffset = CGSize(width: 0, height: 3)
-        viewContainer.layer.shadowRadius = 3
-        viewContainer.layer.shadowOpacity = 0.5
+        containerVStackView.layer.shadowOffset = CGSize(width: 0, height: 5)
+        containerVStackView.layer.shadowRadius = 5
+        containerVStackView.layer.shadowOpacity = 0.3
     }
     
     /// Используется для изменения цвета тени карточки лекарства
@@ -208,39 +334,66 @@ private extension MedicineCell {
     /// нужно вызывать в traitCollectionDidChange для изменение цвета. Так как cgColor
     /// не изменяются сами.
     func setupCardShadowColor() {
-        viewContainer.layer.shadowColor = Colors.cardShadow.cgColor
+        containerVStackView.layer.shadowColor = Colors.cardShadow.cgColor
     }
     
     func addViews() {
-        contentView.addSubview(viewContainer)
-        viewContainer.addSubview(stackMedicineLabels)
-        stackMedicineLabels.addArrangedSubview(nameLabel)
-        stackMedicineLabels.addArrangedSubview(descriptionLabel)
-        stackMedicineLabels.addArrangedSubview(expiryDateLabel)
-        viewContainer.addSubview(actionIcon)
-        viewContainer.addSubview(amountLabel)
+        contentView.addSubview(containerVStackView)
+        
+        containerVStackView.addArrangedSubview(titleContentStackView)
+        titleContentStackView.addArrangedSubview(nameLabel)
+        titleContentStackView.addArrangedSubview(iconImageView)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: titleContentStackView)
+        containerVStackView.addArrangedSubview(separatorTopView)
+        
+        containerVStackView.addArrangedSubview(descriptionLabel)
+        
+        containerVStackView.addArrangedSubview(calculableContentStackView)
+        calculableContentStackView.addArrangedSubview(amountContentStackView)
+        amountContentStackView.addArrangedSubview(staticAmountLabel)
+        amountContentStackView.addArrangedSubview(amountLabel)
+        
+        calculableContentStackView.addArrangedSubview(expiryDateContentStackView)
+        expiryDateContentStackView.addArrangedSubview(staticExpiryDateLabel)
+        expiryDateContentStackView.addArrangedSubview(expiryDateLabel)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: calculableContentStackView)
+        containerVStackView.addArrangedSubview(separatorBottomView)
+        
+        containerVStackView.setCustomSpacing(Indents.elementsPadding, after: separatorBottomView)
+        containerVStackView.addArrangedSubview(bottomContentStackView)
+        bottomContentStackView.addArrangedSubview(actionIcon)
+        bottomContentStackView.addArrangedSubview(editButton)
     }
     
     func setupConstraints() {
         NSLayoutConstraint.activate([
-            viewContainer.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-            viewContainer.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 16),
-            viewContainer.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -16),
-            viewContainer.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
+            containerVStackView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: Indents.edgeMarginPadding),
+            containerVStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: Indents.edgeMarginPadding),
+            containerVStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -Indents.edgeMarginPadding),
+            containerVStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -Indents.edgeMarginPadding),
+                        
+            iconImageView.heightAnchor.constraint(equalToConstant: Constants.iconMedecineSize),
+            iconImageView.widthAnchor.constraint(equalToConstant: Constants.iconMedecineSize),
             
-            stackMedicineLabels.topAnchor.constraint(equalTo: viewContainer.topAnchor, constant: 8),
-            stackMedicineLabels.leadingAnchor.constraint(equalTo: viewContainer.leadingAnchor, constant: 16),
-            stackMedicineLabels.bottomAnchor.constraint(equalTo: viewContainer.bottomAnchor, constant: -8),
+            separatorTopView.heightAnchor.constraint(equalToConstant: Constants.heightBorder),
+            separatorBottomView.heightAnchor.constraint(equalToConstant: Constants.heightBorder),
             
-            actionIcon.widthAnchor.constraint(equalToConstant: 25),
-            actionIcon.heightAnchor.constraint(equalToConstant: 25),
-            actionIcon.topAnchor.constraint(equalTo: viewContainer.topAnchor, constant: 8),
-            actionIcon.leadingAnchor.constraint(equalTo: stackMedicineLabels.trailingAnchor, constant: 8),
-            actionIcon.trailingAnchor.constraint(equalTo: viewContainer.trailingAnchor, constant: -16),
-            actionIcon.bottomAnchor.constraint(lessThanOrEqualTo: amountLabel.topAnchor, constant: -25),
+            actionIcon.widthAnchor.constraint(equalToConstant: Constants.iconActionSize),
             
-            amountLabel.trailingAnchor.constraint(equalTo: viewContainer.trailingAnchor, constant: -16),
-            amountLabel.bottomAnchor.constraint(equalTo: viewContainer.bottomAnchor, constant: -8)
+            editButton.heightAnchor.constraint(equalToConstant: Constants.editButtonHeight),
+            editButton.widthAnchor.constraint(equalToConstant: Constants.editBottonWidth)
         ])
+    }
+}
+
+private extension MedicineCell {
+    struct Constants {
+        static let iconMedecineSize: CGFloat = 48
+        static let iconActionSize: CGFloat = 28
+        static let heightBorder: CGFloat = 2
+        static let editButtonHeight: CGFloat = 50
+        static let editBottonWidth: CGFloat = UIScreen.main.bounds.width * 0.4139 // примерно 178
     }
 }
